base/requirements.txt
#Third party package requirements for OpenEats 2


coreapi-cli
Django==1.10
django-imagekit==3.3
djangorestframework==3.4.6
gunicorn==19.6.0
requests==2.11.1
six==1.10.0
django-filter==0.15.0
django-extensions==1.7.2
Pillow==2.7.0
django-cors-headers==1.3.1
MySQL-python==1.2.5


# Will need these once I start supporting metric mesurments
sympy==0.7.6.1
measurement==1.8.0
# Will neeed to add comments for this
# https://github.com/coddingtonbear/django-measurement

